echoverse-backend/requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
pydantic==2.9.0
python-dotenv==1.0.1
ibm-watsonx-ai==1.3.32
ibm-cos-sdk==2.14.3
requests>=2.32.4
